check len

+++ code/go/0chain.net/blobbercore/reference/dbCollector.go
@@ -34,11 +34,13 @@ func (dc *dbCollector) DeleteRefRecord(ref *Ref) {
 
 func (dc *dbCollector) Finalize(ctx context.Context) error {
 	db := datastore.GetStore().GetTransaction(ctx)
-	err := db.Delete(&dc.deletedRefs).Error
-	if err != nil {
-		return err
+	if len(dc.deletedRefs) > 0 {
+		err := db.Delete(&dc.deletedRefs).Error
+		if err != nil {
+			return err
+		}
 	}
-	err = db.Create(&dc.createdRefs).Error
+	err := db.Create(&dc.createdRefs).Error
 	if err != nil {
 		return err
 	}
